refactor modelio
